separe employee edit or append
